fix(trainee-selection): Moved trainee preferences
